requirements.txt
requests
lxml==4.9.4
cssselect
pytesseract
pillow